chore: 🚚 Rename directory
